More cleanup
Added more comments
Moved telegraph to its own file
clarified mkcpy to a more clear name
